add verification admin user in create, delete, edit and get-list variant use cases, and tests

+++ domain/src/use-cases/variant/delete-variant.ts
@@ -1,21 +1,31 @@
-import { Variant } from "../../entities/variant";
-import { VariantService } from "../../services/variant-service";
-import { DeletePayload } from "../../utils/types/payload";
+import { type User, UserRole } from "../../entities";
+import type { Variant } from "../../entities/variant";
+import type { UserService, VariantService } from "../../services";
+import type { DeletePayload } from "../../utils/types/payload";
 
 interface DeleteVariantDeps {
   variantService: VariantService;
+  userService: UserService;
 }
 
-type DeleteVariantPayload = DeletePayload<Variant>
+type DeleteVariantPayload = DeletePayload<Variant> & {
+  userId: User["id"];
+};
 
 export async function deleteVariant(
-  { variantService }: DeleteVariantDeps,
-  { id }: DeleteVariantPayload
-) : Promise<void | Error> {
+  { variantService, userService }: DeleteVariantDeps,
+  { id, userId }: DeleteVariantPayload
+): Promise<void | Error> {
+  const user = await userService.findById(userId);
+  if (!user) return new Error(`User ${userId} not found`);
+
+  if (user.role !== UserRole.ADMIN) {
+    return new Error(`User is not ${UserRole.ADMIN}`);
+  }
   const foundVariant = await variantService.findById(id);
   if (!foundVariant) {
     return new Error("Variant not found");
   }
 
-  await variantService.delete({id});
+  await variantService.delete({ id });
 }
